% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_sql.R
\name{append_sql}
\alias{append_sql}
\title{Appending data to SQL.}
\usage{
append_sql(data, sql_table, connection, chunksize = 100)
}
\arguments{
\item{data}{The dataframe you want to append to SQL.}

\item{sql_table}{The SQL table you wish to append to. For example, Info.Performance.Perf_Data.}

\item{connection}{The SQL connection. For example, connection_WH or connection_OP.}

\item{chunksize}{Default is 100. This is the number of rows to append at a time. Recommend not changing unless absolutely necessary.}
}
\value{
Rows appended to SQL.
}
\description{
Appending a dataframe to specified SQL table.
}
\examples{
\dontrun{
sql_table <- "Info.AdHoc.MyTableName"
connection_WH <- DBI::dbConnect(odbc::odbc(),
                               Driver = "SQL Server",
                               Server = "oxnetorbit01\\\\tdwh_live",
                               Database = "Info"
                              )
append_sql(data = data,
          sql_table = sql_table,
          connection = connection_WH,
          missing_columns = TRUE
          )
}
}
