% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{email}
\alias{email}
\title{Creating an Email}
\usage{
email(To, Cc, subject, dear = "all", body, table, body2, table2, attachment)
}
\arguments{
\item{To}{A string of email addresses separated by "\verb{;}".}

\item{Cc}{A string of email addresses separated by "\verb{;}".}

\item{subject}{A string.}

\item{dear}{A string. Defaults to "\code{all}".}

\item{body}{A string with paragraphs separated by "\verb{<p>}".}

\item{table}{A table defined in global environment.}

\item{body2}{A string with paragraphs separated by "\verb{<p>}".}

\item{table2}{A table defined in global environment.}

\item{attachment}{A string containing filepath and name using \code{/} instead of
\verb{\\}.}
}
\value{
An email.
}
\description{
Creating an email based on inputs provided. Option to add an attachment.
}
\examples{
\dontrun{
email(To = "address1; address2;", subject = "test",
body = "Para1. <p> Para2.")
}
}
