% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_info_request.R
\name{email_info_request}
\alias{email_info_request}
\title{Creating emails for info requests.}
\usage{
email_info_request(
  To,
  Cc,
  type,
  INF_Number = NULL,
  INF_Name = NULL,
  dear = NULL,
  password = NULL,
  filename = NULL,
  table = NULL
)
}
\arguments{
\item{To}{A string of email addresses separated by "\verb{;}".}

\item{Cc}{A string of email addresses separated by "\verb{;}".}

\item{type}{Either "\code{link}", "\code{password}", "\code{attachment}" or "\code{table}" to
determine email contents.}

\item{INF_Number}{A string.}

\item{INF_Name}{A string.}

\item{dear}{A string.}

\item{password}{A string.}

\item{filename}{A string containing filepath and name of attachment using
\code{/} instead of \verb{\\}.}

\item{table}{A table defined in global environment.}
}
\value{
One of two emails, depending on type specified.
}
\description{
Depending on the type specified, creates an email to send link to data or
password to the requester.
}
\examples{
 \dontrun{email_info_request(type = "password",
To = "email_address",
dear = "email_recipient",
password = "test"
)
}
}
