% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_month.R
\name{financial_month}
\alias{financial_month}
\title{Financial Month.}
\usage{
financial_month(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The financial month, where "1" is April.
}
\description{
Finding the financial month of a date.
}
\examples{
financial_month(Sys.Date())
financial_month(as.Date("2023-03-31"))
financial_month(as.Date("2023-04-01"))
}
