% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_year.R
\name{financial_year}
\alias{financial_year}
\title{Financial Year.}
\usage{
financial_year(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The financial year in YYYY-YYYY format.
}
\description{
Finding the financial year of a date.
}
\examples{
financial_year(Sys.Date())
financial_year(as.Date("2023-03-31"))
financial_year(as.Date("2023-04-01"))
}
