% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_of_month.R
\name{first_of_month}
\alias{first_of_month}
\title{First of the Month}
\usage{
first_of_month(x)
}
\arguments{
\item{x}{A date.}
}
\value{
The first of the month in YYYY-MM-DD format.
}
\description{
Finding the first of the month of a given date.
}
\examples{
first_of_month(Sys.Date())
first_of_month(as.Date("2023-03-28"))
}
