% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_table_excel.R
\name{group_table_excel}
\alias{group_table_excel}
\title{Add One Sheet and Table per Group to Excel.}
\usage{
group_table_excel(
  wb,
  data,
  group,
  range,
  colNames = TRUE,
  column_widths = "auto",
  columns = 1:ncol(data_amend)
)
}
\arguments{
\item{wb}{A string containing the current workbook.}

\item{data}{A dataframe.}

\item{group}{The column you want to group the data into it.}

\item{range}{The range of group values.}

\item{colNames}{Boolean to say whether column names from the dataframe should
be used. Defaults to \code{TRUE}.}

\item{column_widths}{A vector of the desired column widths. Defaults to
"\code{auto}".}

\item{columns}{A vector of the column numbers to apply the specified widths
too. Defaults to all columns.}
}
\value{
Inserts one sheet with data in a table for each group.
}
\description{
Add One Sheet and Table per Group to Excel.
}
\examples{
\dontrun{
wb <- openxlsx::createWorkbook()
data <- cars \%>\% mutate("speed_category" = ifelse(speed <15, "Less than 15",
"Greater than 15"))
group_table_excel(wb = wb, data = data, group = speed_category,
   range = c("Less than 15", "Greater than 15"))
save_excel(wb = wb, filename = "C:/Users/Anya.Ferguson/Desktop/example.xlsx")
}
}
