% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_sunday.R
\name{last_sunday}
\alias{last_sunday}
\title{Last Sunday}
\usage{
last_sunday(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The last Sunday before the given date.
}
\description{
Last Sunday
}
\examples{
last_sunday(Sys.Date())
}
