% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_zero.R
\name{month_zero}
\alias{month_zero}
\title{Adding zero to month.}
\usage{
month_zero(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The month in MM format.
}
\description{
Adding zero to month.
}
\examples{
month_zero(Sys.Date())
month_zero(as.Date("2023-03-28"))
month_zero(as.Date("2023-11-28"))
}
