% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthyear.R
\name{monthyear}
\alias{monthyear}
\title{Getting the Month and Year of a Date}
\usage{
monthyear(date, type)
}
\arguments{
\item{date}{A date.}

\item{type}{Either "\code{abb}", "\code{full}" or "\code{num}" to determine how month should appear.}
}
\value{
A string of the month and year of a date.
}
\description{
Getting the Month and Year of a Date
}
\examples{
monthyear(as.Date("2023-11-28"), "abb")
monthyear(as.Date("2023-11-28"), "full")
monthyear(as.Date("2023-11-28"), "num")
}
