% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_excel.R
\name{plot_excel}
\alias{plot_excel}
\title{Add Plot to Excel}
\usage{
plot_excel(
  wb,
  sheetname,
  plot,
  plot_position = c(2, 1),
  plot_width = 8,
  plot_height = 5
)
}
\arguments{
\item{wb}{A string containing the current workbook.}

\item{sheetname}{A string containing the sheetname.}

\item{plot}{A plot.}

\item{plot_position}{A vector of coordinates of starting column and starting
row. Defaults to column \code{2}, row \code{1}.}

\item{plot_width}{A number for the desired width of plot. Defaults to \code{8}.}

\item{plot_height}{A number for the desired height of plot. Defaults to \code{5}.}
}
\value{
Inserts the plot into the Excel workbook.
}
\description{
Adding a R plot to Excel workbook. Plot can be made using \code{ggplot} or base
\code{R}. Ensure plot exists in the global environment before using the function.
}
\examples{
\dontrun{
plot1 <- plot(cars)
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "example")
plot_excel(sheetname = "example", plot = plot1)
save_excel(wb = wb, filename = "C:/example.xlsx")
}
}
