% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescriptions.R
\name{prescriptions}
\alias{prescriptions}
\title{Prescriptions SQL Query}
\usage{
prescriptions(drugs, startdate, enddate)
}
\arguments{
\item{drugs}{A vector of the drugs you want to search for.}

\item{startdate}{A start date for the search.}

\item{enddate}{An end date for the search.}
}
\value{
A string containing a SQL query to look for the drugs provided within
the specified date range.
}
\description{
To create a base for a prescriptions SQL query that searches columns
CATALOG_CD_DISPLAY, ORDER_MNEMONIC and ORDERED_AS_MNEMONIC in
TDWH_ORH.WareHouse.ODS.d_PrescriptionOrders for all provided drugs in a
specified date range. This query can then be amended before use.
}
\examples{
drugs <- c("Berinert", "Cinryze", "Orladeyo")
startdate <- "2023-06-01"
enddate <- "2023-06-30"
prescriptions(drugs = drugs, startdate = startdate, enddate = enddate)
}
