% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_month.R
\name{previous_month}
\alias{previous_month}
\title{Previous Month}
\usage{
previous_month(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The previous month of that date.
}
\description{
Previous Month
}
\examples{
previous_month(Sys.Date())
previous_month(as.Date("2023-03-28"))
}
