% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{R_Test}
\alias{R_Test}
\title{Title}
\usage{
R_Test(date)
}
\arguments{
\item{date}{Test that it works}
}
\value{
Everything you ever wanted
}
\description{
Title
}
\examples{
 financial_year(Sys.Date())
financial_year(as.Date("2024-01-31"))
financial_year(as.Date("2024-04-01"))
}
