% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_password.R
\name{random_password}
\alias{random_password}
\title{Random Password Generation}
\usage{
random_password(seed = as.character(as.numeric(Sys.time())), length = 6)
}
\arguments{
\item{seed}{Either a number to get the same password each time or a string of
"INF-ABCDE" where ABCDE are numbers. Default uses current timestamp.}

\item{length}{The number of characters in the password. Default is 6.}
}
\value{
A string of length specified of uppercase letters, lowercase letters
and numbers.
}
\description{
Random Password Generation
}
\examples{
random_password(seed = "INF-12345")
random_password(seed = 1, length = 4)
random_password(length = 8)
random_password()
}
