% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Reading .sql files}
\usage{
read_sql(filepath)
}
\arguments{
\item{filepath}{Filepath to the .sql file you wish to read in.}
}
\value{
The contents of the .sql file in one string.
}
\description{
Reading .sql files
}
\examples{
filename <- "C:/SVN/R/Testing read_sql function.sql"
read_sql(filename)
}
