% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavingExcel.R
\name{save_excel}
\alias{save_excel}
\title{Saving an Excel File}
\usage{
save_excel(wb, filename, password)
}
\arguments{
\item{wb}{A string containing the current workbook. Default is "'wb'".}

\item{filename}{A string containing desired filepath and name using \code{/}
instead of \verb{\\}.}

\item{password}{A string containing desired password. If omitted, no password
is assigned.}
}
\value{
Workbook will be saved with filename and password provided.
}
\description{
Saves workbook as Excel file in desired location with desired filename.
Option to add a password.
}
\examples{
\dontrun{
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "example")
save_excel(wb = wb, filename = "C:/Users/Anya.Ferguson/Desktop/example.xlsx")
}
}
