% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmatched_names.R
\name{unmatched_names}
\alias{unmatched_names}
\title{Finding unmatched names}
\usage{
unmatched_names(data, sql_table, connection, filename)
}
\arguments{
\item{data}{The dataframe you want to compare to SQL.}

\item{sql_table}{The SQL table you wish to compare to. For example,
Info.Performance.Perf_Data.}

\item{connection}{The SQL connection. For example, connection_WH or
connection_OP.}

\item{filename}{The filename for the Excel to output to.}
}
\value{
An Excel file.
}
\description{
Comparing a dataframe to a SQL table to create an Excel sheet with the column
names that do not have a match.
}
\examples{
\dontrun{data <- cars
sql_table <- "Info.Performance.Perf_Data"
connection_WH <- DBI::dbConnect(odbc::odbc(),
                                Driver = "SQL Server",
                                Server = "oxnetorbit01\\\\tdwh_live",
                                Database = "Info"
                                )
unmatched_names(data = data,
                sql_table = sql_table,
                connection = connection_WH,
                filename = "C:/example.xlsx"
               )
}
}
