% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_up.R
\name{wrap_up}
\alias{wrap_up}
\title{Wrap up}
\usage{
wrap_up()
}
\value{
Run-time of Rmd.
}
\description{
A function to print the run-time of the Rmd to the html output and to log
start and end datetimes to Info.Adhoc.R_Procedures.
}
\examples{
\dontrun{procedure_start_time <- Sys.Date()
wrap_up()}
}
